AllenAcademy/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
python-dotenv==1.0.1
python-dateutil==2.9.0
